commit of the framework with some little extras
The last days we did intense work so we can do this first commit today.
The files include
- reading and analyzing maps
- finding the path from a to b
- some example-pngs for the streets and buildings
- user-navigation on the map

As we couldnt work on the same file at the same time and it would have been a whole bunch of work to do commits for every single completed function, we decided to locally collect our stuff and upload it as one commit. Further information will come the next few days.

Co-Authored-By: Kikiiii2004 <152853785+Kikiiii2004@users.noreply.github.com>
Co-Authored-By: Elyesa Duru <108876804+freely1967@users.noreply.github.com>